% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BCFM-package.R
\docType{package}
\name{BCFM-package}
\alias{BCFM}
\alias{BCFM-package}
\title{BCFM: Bayesian Covariance Factor Model for Clustering and Factor Analysis}
\description{
Implements Bayesian Covariance Factor Model (BCFM) for clustering and latent factor analysis of multivariate longitudinal data. Provides model selection tools, MCMC sampling via Rcpp, and comprehensive visualization functions for exploring cluster assignments, factor loadings, and model diagnostics.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ategge/BCFM}
  \item Report bugs at \url{https://github.com/ategge/BCFM/issues}
}

}
\author{
\strong{Maintainer}: Meriem Touami \email{meriemt@vt.edu}

Authors:
\itemize{
  \item Allison Tegge \email{ategge@vt.edu}
  \item Marco Ferreira \email{marf@vt.edu}
  \item Shuangshuang Xu \email{xshuangshuang@vt.edu}
}

}
\keyword{internal}
