% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BCFMcpp.R
\name{BCFMcpp}
\alias{BCFMcpp}
\title{Gibbs sampler of BCFM}
\usage{
BCFMcpp(data, model.attributes, hyp.parm, n.iter, every = 1, verbose = FALSE)
}
\arguments{
\item{data}{The dataset}

\item{model.attributes}{Model attributes generated by initialize.model.attributes}

\item{hyp.parm}{Hyperparameters generated by initialize.hyp.parm}

\item{n.iter}{Total number of iterations}

\item{every}{Save every \code{every} iterations}

\item{verbose}{Print the results by every 10th step}
}
\value{
List of Gibbs sampler of parameters described in description
}
\description{
It runs a Gibbs sampler for common factors X, factor loadings B, group mean mu, group covariate Omega, idiosyncratic variances sigma^2, group assignment Z and group probabilities probs. This function uses Rcpp.
}
