% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIC.like.R
\name{BIC.like}
\alias{BIC.like}
\title{BIC-like criterion. Very close to the original BIC method, but this uses the integrated likelihood instead.}
\usage{
\method{BIC}{like}(data.row, Gibbs, model.attributes, burnin = NA)
}
\arguments{
\item{data.row}{The dataset}

\item{Gibbs}{Gibbs sample derived form \code{BCFM} function}

\item{model.attributes}{Model attributes generated by \code{initialize.model.attributes}}

\item{burnin}{Number of burn-in period. If not specified, it uses the first tenths as burn-in period}
}
\value{
The value of Laplace-Metropolis marginal density
}
\description{
It finds a Laplace-Metropolis marginal density of likelihood using posterior mean. It also uses Woodbury lemma for fast calculation
}
