% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getmode.R
\name{getmode}
\alias{getmode}
\title{Get the mode of a vector}
\usage{
getmode(v)
}
\arguments{
\item{v}{The vector to find the mode.}
}
\value{
The mode of \code{v}.
}
\description{
The function returns the mode of a vector.
}
