% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.B.trace.R
\name{ggplot.B.trace}
\alias{ggplot.B.trace}
\title{Trace plot for posterior of factor loadings}
\usage{
\method{ggplot}{B.trace}(
  Gibbs,
  burnin = NA,
  permutation = NA,
  true.val = NA,
  factor.num = 1,
  var_labels = NULL
)
}
\arguments{
\item{Gibbs}{MCMC sample simulated from \code{BCFMR} or \code{BCFMcpp} function}

\item{burnin}{Number of burn-in period. If not specified, no burn-in is removed.}

\item{permutation}{Permutation order vector, if applicable}

\item{true.val}{True values of factor loadings. If not available, NA.}

\item{factor.num}{The index of variable to plot (which variable's loadings across all factors to display). Use NULL to plot all variables.}

\item{var_labels}{Character vector of variable names. If NULL, uses Variable 1, Variable 2, etc.}
}
\value{
A ggplot object showing trace plots
}
\description{
It returns a trace plot of factor loadings, B, showing MCMC convergence
}
