% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.mu.density.R
\name{ggplot.mu.density}
\alias{ggplot.mu.density}
\title{Density of group means mu using ggplot2}
\usage{
\method{ggplot}{mu.density}(
  Gibbs,
  true.val = NA,
  add.legend = FALSE,
  burnin = NA,
  layout.dim = NA,
  main.title = "Posterior Densities of Group Means (mu)",
  x.label = "mu",
  y.label = "Density"
)
}
\arguments{
\item{Gibbs}{The Gibbs sample from BCFM function}

\item{true.val}{The true value of mu, if applicable}

\item{add.legend}{Add legend on extra pane}

\item{burnin}{Number of burn-in period. If not specified, it uses the first tenths as burn-in.}

\item{layout.dim}{Dimension of panes. If not specified, the plots are in one column.}

\item{main.title}{Main title for the entire plot. Default is "Posterior Densities of Group Means (mu)"}

\item{x.label}{X-axis label. Default is "mu"}

\item{y.label}{Y-axis label. Default is "Density"}
}
\value{
A ggplot object (grob from grid.arrange)
}
\description{
The function returns multiple density plots of group mean parameter mu.
}
