% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.Zit.heatmap.R
\name{ggplot.Zit.heatmap}
\alias{ggplot.Zit.heatmap}
\title{A heatmap of group assignments, Z using ggplot2}
\usage{
\method{ggplot}{Zit.heatmap}(
  Gibbs,
  true.val = NA,
  burnin = NA,
  main.title = "Cluster Assignment Heatmap",
  x.label = "Cluster",
  y.label = "Subject-Time"
)
}
\arguments{
\item{Gibbs}{Gibbs sample from \code{BCFM} function}

\item{true.val}{Table of true group assignments, if applicable}

\item{burnin}{Number of burn-in period. If not specified, it uses the first tenths sample as burn-in.}

\item{main.title}{Title of the plot. Default is "Cluster Assignment Heatmap"}

\item{x.label}{X-axis label. Default is "Cluster"}

\item{y.label}{Y-axis label. Default is "Subject-Time"}
}
\value{
A ggplot object
}
\description{
A heatmap of group assignments, Z using ggplot2. It first sorts by the largest group with the most assigned.
}
