% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_B.CI.R
\name{ggplot_B.CI}
\alias{ggplot_B.CI}
\title{Build factor loadings plot}
\usage{
ggplot_B.CI(
  Gibbs,
  true.val = NA,
  burnin = NA,
  permutation = NA,
  main.bool = TRUE,
  var_labels = NULL
)
}
\arguments{
\item{Gibbs}{Result of Gibbs sampler from BCFM function}

\item{true.val}{True values of factor loadings. If not available, NA.}

\item{burnin}{Number of burn-in. If not set, it uses the first tenths as burn-in period.}

\item{permutation}{Permutation of variables. If not set, no permutation.}

\item{main.bool}{Add title of the plots. Default is TRUE.}

\item{var_labels}{Character vector of variable names. If NULL, defaults to Variable 1, Variable 2, etc.}
}
\value{
A ggplot object
}
\description{
The function builds a column-wise plots of factor loadings. The parameters fixed at 1 are displayed with red dashed vertical lines.
}
