% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_probs.trace.R
\name{ggplot_probs.trace}
\alias{ggplot_probs.trace}
\title{Trace plot of probabilities parameter}
\usage{
ggplot_probs.trace(
  Gibbs,
  burnin = NA,
  main.title = "Trace Plot: Cluster Probabilities",
  x.label = "BCFM Iteration (post burn-in)",
  y.label = "Probability",
  cluster_names = NULL
)
}
\arguments{
\item{Gibbs}{Gibbs sample from \code{BCFM} function.}

\item{burnin}{Number of burn-in period. If not specified, it uses the first tenths sample as burn-in.}

\item{main.title}{Title of the plot. Default is "Trace Plot: Cluster Probabilities"}

\item{x.label}{X-axis label. Default is "BCFM Iteration (post burn-in)"}

\item{y.label}{Y-axis label. Default is "Probability"}

\item{cluster_names}{Character vector of cluster names. If NULL, defaults to Cluster 1, Cluster 2, etc.}
}
\value{
A ggplot object showing trace plots
}
\description{
It returns a trace plot of probabilities parameter after burn-in. Different colors represent different groups.
}
