% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_variability.R
\name{ggplot_variability}
\alias{ggplot_variability}
\title{Variability explained by factors}
\usage{
ggplot_variability(
  data,
  nfactors = 5,
  main.title = "Proportion of Variability Explained by Factors",
  x.label = "Number of Factors",
  y.label = "Proportion of Variability"
)
}
\arguments{
\item{data}{The data matrix}

\item{nfactors}{Number of factors to display. Default is 5.}

\item{main.title}{Main title for the plot. Default is "Proportion of Variability Explained by Factors"}

\item{x.label}{X-axis label. Default is "Number of Factors"}

\item{y.label}{Y-axis label. Default is "Proportion of Variability"}
}
\value{
A ggplot object
}
\description{
Plots the proportion of variability explained by each factor based on eigenvalues of the correlation matrix. Useful for determining the number of factors.
}
