% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IC.R
\name{IC}
\alias{IC}
\title{Information Criterion. Very close to the original BIC method, but this uses the integrated likelihood instead.}
\usage{
IC(data.row, Gibbs, model.attributes, cluster.size = 0.05, burnin = NA)
}
\arguments{
\item{data.row}{The dataset}

\item{Gibbs}{Gibbs sample derived form \code{BCFM} function}

\item{model.attributes}{Model attributes generated by \code{initialize.model.attributes}}

\item{cluster.size}{Minimum proportion required for each cluster (default 0.05)}

\item{burnin}{Number of burn-in period. If not specified, it uses the first tenths as burn-in period}
}
\value{
The value of Laplace-Metropolis marginal density
}
\description{
It finds a Laplace-Metropolis marginal density of likelihood using posterior mean. It also uses Woodbury lemma for fast calculation
}
