% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.model.attributes.R
\name{initialize.model.attributes}
\alias{initialize.model.attributes}
\title{Build model attributes from the dataset}
\usage{
initialize.model.attributes(S = 216, times = 5, R = 9, L = 3, G = 4)
}
\arguments{
\item{S}{Number of subjects}

\item{times}{Number of timepoints}

\item{R}{Number of covariates}

\item{L}{Number of factors}

\item{G}{Number of groups}
}
\value{
A list of model attributes
}
\description{
Basic setting and information of the dataset: number of subjects, timepoints, variables, factors and groups.
}
