% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sim.data}
\alias{sim.data}
\title{Simulated BCFM Dataset}
\format{
A list with components:
\describe{
\item{data}{A 3D numeric array (200 x 20 x 5)}
\item{pselected}{Cluster assignments}
\item{X}{Factor scores}
\item{B}{Loading matrix}
}
}
\usage{
sim.data
}
\description{
A simulated dataset for demonstrating the BCFM package functionality.
}
\examples{
data("sim.data", package = "BCFM")
dim(sim.data$data)

}
\keyword{datasets}
