% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BCFM-package.R
\docType{package}
\name{BCFM-package}
\alias{BCFM}
\alias{BCFM-package}
\title{BCFM: Bayesian Clustering Factor Models}
\description{
Implements the Bayesian Clustering Factor Models (BCFM) for simultaneous clustering and latent factor analysis of multivariate longitudinal data. The model accounts for within-cluster dependence through shared latent factors while allowing heterogeneity across clusters, enabling flexible covariance modeling in high-dimensional settings. Inference is performed using Markov chain Monte Carlo (MCMC) methods with computationally intensive steps implemented via 'Rcpp'. Model selection and visualization tools are provided. The methodology is described in Shin, Ferreira, and Tegge (2018) \doi{10.1002/sim.70350}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ategge/BCFM}
  \item Report bugs at \url{https://github.com/ategge/BCFM/issues}
}

}
\author{
\strong{Maintainer}: Meriem Touami \email{meriemt@vt.edu}

Authors:
\itemize{
  \item Allison Tegge \email{ategge@vt.edu}
  \item Marco Ferreira \email{marf@vt.edu}
  \item Hwasoo Shin \email{hshin2@hfhs.org}
}

}
\keyword{internal}
